Use timedelta to check for expiration
'seconds' in a timedelta object are always positive, so it was not the
best idea to use it to check whether tokens have expired, as tokens
expired longer than the renewal interval are never found to be expired.

Fix this by making the renewal interval a timedelta and compare it with
the timedelta between the expiration time and the current time.

Add two new tests to cover the case when tokens are already expired.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>