Run gunicorn with Python
Instead of calling 'gunicorn', run it through Python.

Do this, b/c the entrypoint script installed by gunicorns RPM package,
uses the '-s' flag for the Python interpreter and so it doesn't add to
'sys.path' the user site directory, which causes modules installed with
'pip' (like sentry-sdk) not to be found.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>