Get ready to upgrade to PyGithub 1.58
[PyGithub 1.58] introduces proper support for GitHub App auth. The old
interfaces are still available, but their behaviour changed:
get_installation() will raise a GithubException if the GitHub App is not
installed for a given repository, instead of returning None.

Make sure to catch this exception so that things don't brake when
PyGithub 1.58 lands (probably when upgrading the base image to Fedora
Linux 38).

[PyGithub 1.58]: https://github.com/PyGithub/PyGithub/releases/tag/v1.58.0

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>